% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_state_info.R
\name{decode_LOH}
\alias{decode_LOH}
\title{Decode the Titan State To Give Copy number and State Name.}
\usage{
decode_LOH(G, symmetric = TRUE)
}
\arguments{
\item{G}{titan state}

\item{symmetric}{Boolean flag to indicate whether "similar" states should be collapsed}
}
\description{
Decode the Titan State To Give Copy number and State Name.
}
\author{
Gavin Ha
\url{https://github.com/gavinha/TitanCNA/blob/master/R/utils.R}
}

