% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_gene_titan_seg_overlap.R
\name{find_gene_titan_seg_overlap}
\alias{find_gene_titan_seg_overlap}
\title{Get Gene Overlaps with TitanCNA segments}
\usage{
find_gene_titan_seg_overlap(titan.seg.df, gene.annot.df)
}
\arguments{
\item{titan.seg.gr}{GRanges object of the TitanCNA seg}

\item{gene.annot.gr}{GRanges object of the genes}
}
\value{
data.frame that contains the overlap of the input gene annotation
 with the input segment data.
}
\description{
Get Gene Overlaps with TitanCNA segments
}

