% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_gene_titan_seg_overlap.R
\name{get_gene_titan_seg_overlap}
\alias{get_gene_titan_seg_overlap}
\title{Get Gene Overlaps with TitanCNA segments}
\usage{
get_gene_titan_seg_overlap(segs.df, gene.annot.df)
}
\arguments{
\item{segs.df}{data.frame of the TitanCNA segs file loaded from
\code{oad_titan_seg}}

\item{gene.annot.df}{}
}
\value{
data.frame that contains the overlap of the input gene annotation
 with the input segment data.
}
\description{
Get Gene Overlaps with TitanCNA segments
}

