% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_state_info.R
\name{get_state_info}
\alias{get_state_info}
\title{Get Titan State Information}
\usage{
get_state_info()
}
\value{
A named list containing various details "categories" of state
  information
}
\description{
Returns information about the Titan states in a list structure
}

