% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_titan_seg.R
\name{load_titan_seg}
\alias{load_titan_seg}
\title{Load the TitanCNA Seg File}
\usage{
load_titan_seg(titan.seg.file)
}
\arguments{
\item{titan.seg.file}{Path to TitanCNA seg file. This should be the file
generated from the createTITANsegmentfiles.pl script provided in the
TitanCNA download}
}
\value{
data.frame of the TitanCNA seg file
}
\description{
Load the TitanCNA Seg File
}

