% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt_gene_call_mat.R
\name{melt_gene_call_mat}
\alias{melt_gene_call_mat}
\title{Melts the Titan Gene Call Matrix into Tidy Format}
\usage{
melt_gene_call_mat(in.gene.call.mat, titan.state.info)
}
\arguments{
\item{titan.state.info}{Output from the get_state_info() function}

\item{in.gene.call.matrix}{The gene call matrix}
}
\value{
A 4 column data.frame of the Titan gene call data.
}
\description{
This function melts the Titan gene call matrix into a tidy format. It also
adds an additional column called stateSummary. This column represents the
the states summarized into broad states as specified in the titan.state.info
variable
}

