% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titanCNAutils.R
\docType{package}
\name{titanCNAutils}
\alias{titanCNAutils}
\alias{titanCNAutils-package}
\title{titanCNAutils: An R package for working TitanCNA Results}
\description{
titanCNAutils provides various utility functions for post-processing
TitanCNA results.
}

