% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_overlap_seg_with_mask_ind.R
\name{get_overlap_seg_with_mask_ind}
\alias{get_overlap_seg_with_mask_ind}
\title{Find Overlapping Segments with a CNV Mask}
\usage{
get_overlap_seg_with_mask_ind(segs.df, cnv.mask.df, overlap.prop = 0.25)
}
\arguments{
\item{segs.df}{Titan segments data.frame}

\item{cnv.mask.df}{CNV mask in a data.frame. Must contain the columns: chr,
start, end}

\item{overlap.prop}{Proportion that the TitanCNA segment must overlap with
cnv mask segment in order to be considered overlapping}
}
\value{
Vector containing the row number in the segs.df that overlaps with
 any cnv.mask segment
}
\description{
Given a TitanCNA segment data.frame and a CNV mask segment, this function
will find whether any TitanCNA segment overlaps, with any significant degree
according to the overlap.prop parameter, with a CNV mask segment. It returns
the row number of any segment that does.
}

